+++ e2e/components/upload/upload-surveillance/upload-surveillance.spec.js
@@ -3,7 +3,10 @@ import LoginComponent from '../../login/login.po';
 import Hooks from '../../../utilities/hooks';
 import ToastComponent from '../../toast/toast.po';
 
-let hooks, loginComponent, toast, uploadSurveillanceComponent;
+let hooks;
+let loginComponent;
+let toast;
+let uploadSurveillanceComponent;
 
 beforeEach(async () => {
   uploadSurveillanceComponent = new UploadSurveillanceComponent();
@@ -14,11 +17,11 @@ beforeEach(async () => {
 });
 
 describe('when uploading a surveillance activity as ONC-ACB', () => {
-  beforeEach(function () {
+  beforeEach(() => {
     loginComponent.logIn('drummond');
   });
 
-  afterEach(function () {
+  afterEach(() => {
     hooks.waitForSpinnerToDisappear();
     toast.clearAllToast();
     loginComponent.logOut();
@@ -26,13 +29,13 @@ describe('when uploading a surveillance activity as ONC-ACB', () => {
 
   it('can upload valid file format of surveillance activity file', () => {
     uploadSurveillanceComponent.uploadSurveillance('../../../resources/surveillance/SAQA1.csv');
-    browser.waitUntil( () => toast.toastTitle.isDisplayed());
+    browser.waitUntil(() => toast.toastTitle.isDisplayed());
     expect(toast.toastTitle.getText()).toBe('Success');
   });
 
   it('can\'t upload invalid file format of surveillance activity file', () => {
     uploadSurveillanceComponent.uploadSurveillance('../../../resources/surveillance/SurveillanceInvalid.csv');
-    browser.waitUntil( () => toast.toastTitle.isDisplayed());
+    browser.waitUntil(() => toast.toastTitle.isDisplayed());
     expect(toast.toastTitle.getText()).toBe('Error');
   });
 });