+++ src/app/components/products/product/edit.component.js
@@ -1,4 +1,4 @@
-export const ProductEditComponent = {
+const ProductEditComponent = {
   templateUrl: 'chpl.components/products/product/edit.html',
   bindings: {
     product: '<',
@@ -8,41 +8,42 @@ export const ProductEditComponent = {
     takeAction: '&',
   },
   controller: class ProductEditComponent {
-    constructor ($log, networkService) {
+    constructor($log, networkService) {
       'ngInject';
+
       this.$log = $log;
       this.networkService = networkService;
       this.mergeOptions = {};
     }
 
-    $onInit () {
-      let that = this;
-      this.networkService.getDevelopers(true).then(response => {
+    $onInit() {
+      const that = this;
+      this.networkService.getDevelopers(true).then((response) => {
         that.developers = response.developers
-          .map(d => {
-            d.displayName = d.name + (d.deleted ? ' - deleted' : ' - active');
-            return d;
-          })
-          .sort((a, b) => a.name < b.name ? -1 : a.name > b.name ? 1 : 0);
+          .map((d) => ({
+            ...d,
+            displayName: d.name + (d.deleted ? ' - deleted' : ' - active'),
+          }))
+          .sort((a, b) => (a.name < b.name ? -1 : 0));
         if (that.developers && that.product) {
-          that.currentOwner = that.developers.filter(d => d.developerId === that.product.owner.developerId)[0];
+          [that.currentOwner] = that.developers.filter((d) => d.developerId === that.product.owner.developerId);
         }
       });
     }
 
-    $onChanges (changes) {
+    $onChanges(changes) {
       if (changes.product && changes.product.currentValue) {
         this.product = angular.copy(changes.product.currentValue);
         this.product.ownerHistory = this.product.ownerHistory
-          .filter(o => o.transferDate)
+          .filter((o) => o.transferDate)
           .concat({
             developer: this.product.owner,
             transferDate: undefined,
           })
-          .map(d => {
-            d.displayName = d.name + (d.deleted ? ' - deleted' : ' - active');
-            return d;
-          })
+          .map((d) => ({
+            ...d,
+            displayName: d.name + (d.deleted ? ' - deleted' : ' - active'),
+          }))
           .sort((a, b) => {
             if (a.transferDate && b.transferDate) {
               return b.transferDate - a.transferDate;
@@ -66,11 +67,11 @@ export const ProductEditComponent = {
       }
     }
 
-    cancel () {
-      this.takeAction({action: 'cancel'});
+    cancel() {
+      this.takeAction({ action: 'cancel' });
     }
 
-    doneAddingOwner () {
+    doneAddingOwner() {
       this.newOwner = undefined;
       this.newTransferDate = undefined;
       this.addingOwner = false;
@@ -78,13 +79,13 @@ export const ProductEditComponent = {
       this.generateErrorMessages();
     }
 
-    editContact (contact) {
+    editContact(contact) {
       this.product.contact = angular.copy(contact);
       this.generateErrorMessages();
     }
 
-    generateErrorMessages () {
-      let messages = [];
+    generateErrorMessages() {
+      const messages = [];
       if (this.showFormErrors) {
         if (this.isMerging && (!this.mergingProducts || this.mergingProducts.length === 0)) {
           messages.push('At least one other Product must be selected to merge');
@@ -94,7 +95,7 @@ export const ProductEditComponent = {
         this.product.ownerHistory.forEach((o, idx, arr) => {
           if (idx > 0) {
             if (arr[idx].developer.name === arr[idx - 1].developer.name) {
-              messages.push('Product cannot transfer from Developer "' + arr[idx].developer.name + '" to the same Developer');
+              messages.push(`Product cannot transfer from Developer "${arr[idx].developer.name}" to the same Developer`);
             }
           }
         });
@@ -102,26 +103,26 @@ export const ProductEditComponent = {
       this.errorMessages = messages;
     }
 
-    isValid () {
+    isValid() {
       return this.errorMessages.length === 0 // business logic rules
-                && this.form.$valid; // form validation
+        && this.form.$valid; // form validation
     }
 
-    removeOwner (owner) {
+    removeOwner(owner) {
       this.product.ownerHistory = this.product.ownerHistory
-        .filter(o => (!(o.developer.developerId === owner.developer.developerId && o.transferDate === owner.transferDate)));
+        .filter((o) => (!(o.developer.developerId === owner.developer.developerId && o.transferDate === owner.transferDate)));
       this.generateErrorMessages();
     }
 
-    save () {
-      let request = angular.copy(this.product);
+    save() {
+      const request = angular.copy(this.product);
       request.owner = request.ownerHistory[0].developer;
-      request.ownerHistory = request.ownerHistory.filter(o => o.transferDate);
-      this.takeAction({action: 'edit', data: request});
+      request.ownerHistory = request.ownerHistory.filter((o) => o.transferDate);
+      this.takeAction({ action: 'edit', data: request });
     }
 
-    saveNewOwner () {
-      let time = this.newTransferDate ? this.newTransferDate.getTime() : undefined;
+    saveNewOwner() {
+      const time = this.newTransferDate ? this.newTransferDate.getTime() : undefined;
       this.product.ownerHistory = this.product.ownerHistory
         .concat({
           developer: this.newOwner,
@@ -136,29 +137,28 @@ export const ProductEditComponent = {
       this.doneAddingOwner();
     }
 
-    takeActionBarAction (action) {
+    takeActionBarAction(action) {
       switch (action) {
-      case 'cancel':
-        this.cancel();
-        break;
-      case 'mouseover':
-        this.showFormErrors = true;
-        this.generateErrorMessages();
-        break;
-      case 'save':
-        this.save();
-        break;
-                //no default
+        case 'cancel':
+          this.cancel();
+          break;
+        case 'mouseover':
+          this.showFormErrors = true;
+          this.generateErrorMessages();
+          break;
+        case 'save':
+          this.save();
+          break;
+          // no default
       }
     }
 
-    updateCurrentOwner () {
-      this.product.ownerHistory.forEach(owner => {
-        if (!owner.transferDate) {
-          owner.transferDate = (new Date()).getTime();
-        }
-      });
+    updateCurrentOwner() {
       this.product.ownerHistory = this.product.ownerHistory
+        .map((owner) => ({
+          ...owner,
+          transferDate: owner.transferDate || Date.now(),
+        }))
         .concat({
           developer: this.currentOwner,
           transferDate: undefined,
@@ -173,51 +173,22 @@ export const ProductEditComponent = {
     }
 
     /*
-         * Pill generation
-         */
-    generateMergeOptions () {
+     * Pill generation
+     */
+    generateMergeOptions() {
       this.mergeOptions = {
-        name: Array.from(new Set([this.productBackup.name].concat(this.mergingProducts.map(p => p.name)))),
+        name: Array.from(new Set([this.productBackup.name].concat(this.mergingProducts.map((p) => p.name)))),
       };
-      this.contactOptions = {
-        fullName: [],
-        title: [],
-        email: [],
-        phoneNumber: [],
-      };
-      this.fillMergeOptionByProduct(this.productBackup);
-      this.mergingProducts.forEach(p => this.fillMergeOptionByProduct(p));
-      this.contactOptions.fullName = Array.from(new Set(this.contactOptions.fullName));
-      this.contactOptions.title = Array.from(new Set(this.contactOptions.title));
-      this.contactOptions.email = Array.from(new Set(this.contactOptions.email));
-      this.contactOptions.phoneNumber = Array.from(new Set(this.contactOptions.phoneNumber));
     }
 
-    fillMergeOptionByProduct (product) {
-      if (product.contact) {
-        if (product.contact.fullName) {
-          this.contactOptions.fullName.push(product.contact.fullName);
-        }
-        if (product.contact.title) {
-          this.contactOptions.title.push(product.contact.title);
-        }
-        if (product.contact.email) {
-          this.contactOptions.email.push(product.contact.email);
-        }
-        if (product.contact.phoneNumber) {
-          this.contactOptions.phoneNumber.push(product.contact.phoneNumber);
-        }
-      }
-    }
-
-    getDifferences (predicate) {
-      if (!this.product || !this.mergeOptions[predicate]) { return; }
+    getDifferences(predicate) {
+      if (!this.product || !this.mergeOptions[predicate]) { return []; }
       return this.mergeOptions[predicate]
-        .filter(e => e && e.length > 0 && e !== this.product[predicate])
-        .sort((a, b) => a < b ? -1 : a > b ? 1 : 0);
+        .filter((e) => e && e.length > 0 && e !== this.product[predicate])
+        .sort((a, b) => (a < b ? -1 : 1));
     }
 
-    selectDifference (predicate, value) {
+    selectDifference(predicate, value) {
       this.product[predicate] = value;
     }
   },
@@ -225,3 +196,5 @@ export const ProductEditComponent = {
 
 angular.module('chpl.components')
   .component('chplProductEdit', ProductEditComponent);
+
+export default ProductEditComponent;